Moved Functions into new files & updated package.json
