+++ .gitignore
@@ -1,2 +1,3 @@
 evidence.txt
 js/giphy.js
+node_modules
