+++ index.html
@@ -12,7 +12,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Ubuntu+Mono:ital,wght@0,400;0,700;1,400;1,700&display=swap"
         rel="stylesheet">
 
-    <script src="js/load-cat.js"></script>
+    <script src="./main.js"></script>
 </head>
 
 <body>
@@ -126,6 +126,8 @@
                 </aside>
             </section>
         </article>
+
+        <section id="gif-container"></section>
     </main>
     <footer>
         <span aria-hidden="true" id="footer-cat-carrier" class="cat-carrier"></span>
