--- js/load-cat.js
@@ -1,20 +1,0 @@
-window.onload = () => {
-    const footerCat = `
-         ,_     _
-         |\\_,-~/
-         / _  _ |    ,--.
-        (  @  @ )   / ,-'
-         \\  _T_/-._( (
-         /         \`. \\
-        |         _  \\ |
-         \\ \\ ,  /      |
-          || |-_\\__   /
-         ((_/\`(____,-'
-    `;
-    loadFooterCat(footerCat);
-}
-
-function loadFooterCat(footerCat) {
-    document.getElementById("footer-cat-carrier").innerHTML = footerCat;
-}
-
