+++ js/loadFooterCat.js
@@ -1,0 +1,15 @@
+export function loadFooterCat() {
+  const footerCat = `
+    ,_     _
+    |\\_,-~/
+    / _  _ |    ,--.
+   (  @  @ )   / ,-'
+    \\  _T_/-._( (
+    /         \`. \\
+   |         _  \\ |
+    \\ \\ ,  /      |
+     || |-_\\__   /
+    ((_/\`(____,-'
+`;
+  document.getElementById("footer-cat-carrier").innerHTML = footerCat;
+}
